+++ x/ibcdex/types/account_test.go
@@ -4,15 +4,15 @@ import "github.com/tendermint/interchange/x/interchange/types"
 
 func GenLocalAccount() types.Account {
 	return types.Account{
-		Address: GenAddress(),
+		Address: types.GenAddress(),
 		Chain:   "",
 	}
 }
 
 func GenRemoteAccount() types.Account {
 	return types.Account{
-		Address: GenAddress(),
-		Chain:   GenString(10),
+		Address: types.GenAddress(),
+		Chain:   types.GenString(10),
 	}
 }
