--- x/interchange/types/account.go
@@ -1,6 +1,0 @@
-package types
-
-type Account struct {
-	Address string // sdk.AccAddress
-	Chain   string // "" if actual chain, "portID-channelID" if distant chain
-}
